+++ resources/lang/en/route-descriptions.php
@@ -59,7 +59,7 @@ return [
     'summary_GET_months' => 'Return the monthly summary (sum of items) for the selected resource and year',
     'summary_GET_month' => 'Return the monthly summary (sum of items) for the selected resource, year and month',
 
-    'request_GET_log' => 'Return the requests log, read requests',
+    'request_GET_access-log' => 'Return the access log, read requests',
     'request_GET_log_monthly_requests' => 'Return the monthly summary of read requests',
     'request_GET_error_log' => 'Return the error log',
     'request_POST' => 'Create an error log report',